refactor: modularize LineConfidenceChart by integrating MainChart and OverviewChart components
- Replaced the existing chart rendering logic in LineConfidenceChart with MainChart and OverviewChart components for improved code organization and readability.
- Enhanced maintainability by separating concerns between main and overview chart functionalities.
- Updated props passed to the new components to ensure consistent data handling and rendering.
